requirements.txt
azure-identity
aiohttp
ipykernel
dotenv
requests
msgraph-sdk
azure-search-documents==11.6.0b12
fastmcp
click